LayoutTests/fast/css/matrix-stringifier.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
["DOMMatrix", "DOMMatrixReadOnly", "WebKitCSSMatrix"].forEach(constr => {
    const prefix = `${constr} stringifier:`;
    test(() => {
        const matrix = new self[constr]();
        assert_equals(String(matrix), "matrix(1, 0, 0, 1, 0, 0)");
    }, `${prefix} identity (2d)`);

    test(() => {
        const matrix = new self[constr]("matrix(1, 0, 0, 1, 0, 0)");
        assert_equals(String(matrix), "matrix(1, 0, 0, 1, 0, 0)");
    }, `${prefix} integer 2d`);
 
    test(() => {
        const matrix = new self[constr]("matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1)");
        assert_equals(String(matrix), "matrix(1, 0, 0, 1, 0, 0)");
    }, `${prefix} integer 3d`);

    test(() => {
        const matrix = new self[constr]("matrix(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)");
        assert_equals(String(matrix), "matrix(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)");
    }, `${prefix} floating point 2d`)

    test(() => {
        const matrix = new self[constr]("matrix3d(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7)");
        assert_equals(String(matrix), "matrix3d(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7)");
    }, `${prefix} floating point 3d`);

    [NaN, Infinity, -Infinity].forEach(num => {
        test(() => {
            const matrix = new self[constr]();
            matrix.a = num;
            assert_throws("InvalidStateError", () => String(matrix));
        }, `${prefix} ${num}`);
    });

    test(() => {
        assert_true(Object.getOwnPropertyDescriptor(self[constr].prototype, "toString").enumerable);
    }, `${constr}.toString should be enumerable`);
});
</script>
</body>
</html>
